+++ R/role_playing.R
@@ -184,6 +184,30 @@ simplify_rpd_dimensions <- function(db, names) {
 }
 
 
+
+#' From rpd dimensions, leave only contained in vector of names.
+#'
+#' @param db A `star_database` object.
+#' @param names A vector of strings, dimension names.
+#'
+#' @return A list of vectors of dimension names.
+#'
+#' @keywords internal
+filter_rpd_dimensions <- function(db, names) {
+  rpd <- list()
+  rpd_names <- NULL
+  for (i in seq_along(db$rpd)) {
+    db$rpd[[i]] <- intersect(db$rpd[[i]], names)
+    if (length(db$rpd[[i]]) > 1) {
+      rpd <- c(rpd, list(db$rpd[[i]]))
+      rpd_names <- c(rpd_names, db$rpd[[i]][1])
+    }
+  }
+  names(rpd) <- rpd_names
+  rpd
+}
+
+
 #' Share the given dimensions in the database
 #'
 #' @param db `star_database` object.